requirements.txt
pydantic==1.10.8
PyYAML==6.0
superagi-tools
autospark_kit==0.0.1
pyartifactory